+++ indico/modules/search/client/js/components/SearchBox.jsx
@@ -6,118 +6,111 @@
 // LICENSE file for more details.
 
 import PropTypes from 'prop-types';
-import React, {useState} from 'react';
-import {Checkbox, Form, Icon, Label, Search} from 'semantic-ui-react';
+import React, {useRef, useState} from 'react';
 
-import {Translate} from 'indico/react/i18n';
+import {Checkbox} from 'indico/react/components';
+import {useVerticalArrowNavigation} from 'indico/react/hooks';
+import {Param, Translate} from 'indico/react/i18n';
 
 import './SearchBox.module.scss';
 
-const renderResult = ({label}, keyword) => (
-  <div styleName="search-type-option">
-    <span>
-      <Icon name="search" />
-      {keyword}
-    </span>
-    <Label content={label} size="small" />
-  </div>
-);
+function CategoryOption({keyword, scopeLabel, formAction}) {
+  return (
+    <li styleName="search-type-option">
+      <button type="submit" formAction={formAction} data-stop>
+        <Translate>
+          Find "<Param name="keyword" value={keyword} />" in{' '}
+          <Param wrapper={<span styleName="search-scope" />} name="scope" value={scopeLabel} />
+        </Translate>
+      </button>
+    </li>
+  );
+}
 
-export default function SearchBox({onSearch, category, isAdmin}) {
-  const [keyword, setKeyword] = useState('');
-  const [adminOverrideEnabled, setAdminOverrideEnabled] = useState(false);
-  const options =
-    category && !category.isRoot
-      ? [
-          {
-            value: 'category',
-            label: Translate.string('In this category ↵'),
-            title: category.title,
-            renderer: props => renderResult(props, keyword),
-          },
-          {
-            value: 'global',
-            label: Translate.string('All Indico ↵'),
-            title: 'Global',
-            renderer: props => renderResult(props, keyword),
-          },
-        ]
-      : [];
-  if (isAdmin) {
-    options.push({
-      value: 'enableAdminOverride',
-      title: 'Enable admin override',
-      className: 'cursor-default',
-      onClick: () => {},
-      // eslint-disable-next-line react/display-name
-      renderer: () => (
-        <div styleName="search-type-option">
-          <Label content={Translate.string('ADMIN')} size="small" color="red" />
-          <Checkbox
-            id="checkbox-admin-search"
-            styleName="checkbox-admin-search"
-            label={Translate.string('Skip access checks')}
-            checked={adminOverrideEnabled}
-            onChange={e => {
-              e.stopPropagation();
-              setAdminOverrideEnabled(!adminOverrideEnabled);
-            }}
-          />
-        </div>
-      ),
-    });
-  }
+CategoryOption.propTypes = {
+  keyword: PropTypes.string.isRequired,
+  scopeLabel: PropTypes.string.isRequired,
+  formAction: PropTypes.string,
+};
 
-  const handleSearchChange = event => {
-    setKeyword(event.target.value);
-  };
+CategoryOption.defaultProps = {
+  formAction: undefined,
+};
 
-  const handleSubmit = () => {
-    if (keyword.trim()) {
-      onSearch(keyword.trim(), category ? category.isRoot : true, adminOverrideEnabled);
-    }
-  };
-
-  const handleResultSelect = (e, data) => {
-    if (data.result.value === 'enableAdminOverride') {
-      setAdminOverrideEnabled(!adminOverrideEnabled);
-    } else if (keyword.trim()) {
-      onSearch(keyword.trim(), data.result.value === 'global', adminOverrideEnabled);
-    }
+function AdminOverrideOption() {
+  // Keeps focus within the search box widget as we rely on
+  // focus state to keep the popup open.
+  const focusCheckbox = evt => {
+    const label = evt.currentTarget;
+    // Focusout fires *after* this event and it would annul the
+    // `.focus()` call below, so we need to wait for it.
+    requestAnimationFrame(() => label.querySelector('input').focus());
   };
 
   return (
-    <Form onSubmit={handleSubmit}>
-      <label styleName="search-field">
-        <Translate as="span">Search Indico</Translate>
-        <Search
-          input={{fluid: true}}
-          placeholder={Translate.string('Enter your search term')}
-          results={options}
+    <li styleName="admin-check">
+      <label onPointerDown={focusCheckbox}>
+        <Translate as="span" styleName="admin-label">
+          Admin
+        </Translate>
+        <span styleName="admin-check">
+          <Checkbox name="admin_override_enabled" value="true" data-stop />
+          <Translate as="span">Skip access checks</Translate>
+        </span>
+      </label>
+    </li>
+  );
+}
+
+export default function SearchBox({options, defaultUrl, isAdmin}) {
+  const [keyword, setKeyword] = useState('');
+  const searchOptions = [];
+  const formRef = useRef();
+  const inputRef = useRef();
+
+  if (keyword) {
+    for (const opt of options) {
+      searchOptions.push(<CategoryOption key={opt.scopeLabel} keyword={keyword} {...opt} />);
+    }
+  }
+
+  if (isAdmin) {
+    searchOptions.push(<AdminOverrideOption key="override" />);
+  }
+
+  const handleKeywordChange = ev => {
+    setKeyword(ev.target.value.trim());
+  };
+
+  useVerticalArrowNavigation(formRef, inputRef);
+
+  return (
+    <form ref={formRef} action={defaultUrl} styleName="search-box" role="application">
+      <label styleName="field">
+        <Translate as="span">Search (use the arrow keys to select search options).</Translate>
+        <input
+          ref={inputRef}
+          type="text"
+          name="q"
+          onChange={handleKeywordChange}
           value={keyword}
-          showNoResults={false}
-          open={!!keyword}
-          onResultSelect={handleResultSelect}
-          onSearchChange={handleSearchChange}
-          fluid
-          // only select the first result if there is more than 1 (means we are in a category)
-          selectFirstResult={options.length > 1}
+          autoComplete="off"
+          required
+          data-stop
         />
       </label>
-    </Form>
+      <div styleName="options">{searchOptions}</div>
+    </form>
   );
 }
 
 SearchBox.propTypes = {
-  onSearch: PropTypes.func.isRequired,
-  category: PropTypes.shape({
-    id: PropTypes.number.isRequired,
-    title: PropTypes.string.isRequired,
-    isRoot: PropTypes.bool.isRequired,
-  }),
+  defaultUrl: PropTypes.string.isRequired,
+  options: PropTypes.arrayOf(
+    PropTypes.shape({
+      scopeLabel: PropTypes.string.isRequired,
+      formAction: PropTypes.string,
+    })
+  ).isRequired,
   isAdmin: PropTypes.bool.isRequired,
-};
-
-SearchBox.defaultProps = {
-  category: null,
 };